requirements.txt
asgiref==3.5.2
config==0.5.1
defusedxml==0.7.1
diff-match-patch==20200713
Django==4.0.5
django-cors-headers==3.13.0 
django-crispy-forms==1.14.0
django-import-export==2.8.0
django-nested-admin==3.4.0
djangorestframework==3.13.1
et-xmlfile==1.1.0
install==1.3.5
MarkupPy==1.14
odfpy==1.4.1
openpyxl==3.0.10
python-monkey-business==1.0.0
pytz==2022.1
PyYAML==6.0
six==1.16.0
sqlparse==0.4.2
tablib==3.2.1
uWSGI==2.0.20
xlrd==2.0.1
XlsxWriter==3.0.3
xlwt==1.3.0
